refactor SafeActionFormModal
